Fixed DeconditionedDistribution
Now the case of mixed continuous/discrete marginals for a conditioning distribution with dependence is excluded. Either the marginals are independent, or they are of the same kind.
The discretization is no more based on WeightedExperiment as it involves a significant overhead in the constructor. It is done using lower level classes to speed up things.